Update test3.py
changing limits and adding some more comments.
Crash maybe because of not finding data for that star, so ignore the star for which it is not finding data.
Also check with 2nd obj list to (1) skip ones with the asteroseismic detections, and prioritize those with a rotation period. Look at some comments for further suggestions.